daemon/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cloud-drive-sync"
version = "0.1.0"
description = "Multi-cloud bidirectional file sync daemon for Linux"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.2.0",
    "watchdog>=4.0.0",
    "aiofiles>=23.2.0",
    "tomli-w>=1.0.0",
    "click>=8.1.0",
    "cryptography>=41.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dropbox = ["dropbox>=12.0.0"]
onedrive = ["msgraph-sdk>=1.0.0", "azure-identity>=1.15.0"]
nextcloud = ["nc-py-api>=0.17.0"]
box = ["box-sdk-gen>=1.0.0"]
all-providers = ["cloud-drive-sync[dropbox,onedrive,nextcloud,box]"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
cloud-drive-sync-daemon = "cloud_drive_sync.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cloud_drive_sync"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: integration tests using demo mode (no real Google credentials)",
]

[tool.ruff]
target-version = "py312"
line-length = 100
